#117 Added support for multiline strings
