+++ pkg/compiler/compiler_for_ternary.go
