+++ pkg/parser/antlr/FqlLexer.tokens
@@ -57,8 +57,9 @@ In=56
 Param=57
 Identifier=58
 StringLiteral=59
-IntegerLiteral=60
-FloatLiteral=61
+TemplateStringLiteral=60
+IntegerLiteral=61
+FloatLiteral=62
 ':'=5
 ';'=6
 '.'=7
