+++ pkg/parser/antlr/FqlParser.g4
@@ -154,6 +154,7 @@ booleanLiteral
 
 stringLiteral
     : StringLiteral
+    | TemplateStringLiteral
     ;
 
 integerLiteral
